move types tests to back to be a "real" pkg/test
